requirements.txt
polygon-api-client>=1.14.0
pandas>=2.0.0
numpy>=1.25.0
python-json-logger>=2.0.7
tqdm>=4.66.0
requests>=2.31.0
ta>=0.10.2
xgboost>=2.0.0
scikit-learn>=1.4.0
